app/backend/ai-analyzer/pyproject.toml
[tool.poetry]
name = "ai-analyzer"
version = "0.1.0"
description = ""
authors = ["Fred Bone <bone_fred2007@yahoo.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.2"
langchain-experimental = "^0.3.2"
openai = "^1.52.0"
langchain-openai = "^0.2.3"
psycopg2 = "^2.9.10"
jinja2 = "^3.1.4"
transformers = "^4.48.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
